Broke AS4100_5 into 2x files (moment & shear) Added additional test file and did some sorting.
