+++ AS4100 Modules/Verification/UT_AS4100_5.py
@@ -58,45 +58,3 @@ class Test_UT_AS4100_5(TestCase):
         '''
 
         self.fail()
-
-    def test_s5_11_4_V_w_Generic(self):
-        '''
-        Test the s5_11_4_V_w_Generic method.
-        '''
-
-        self.fail()
-
-    def test_s5_11_4_V_w_Weldlimited(self):
-        '''
-        Test the s5_11_4_V_w_Weldlimited method.
-        '''
-
-        self.fail()
-
-    def test_s5_11_4_V_w_CHS(self):
-        '''
-        Test the s5_11_4_V_w_CHS method.
-        '''
-
-        self.fail()
-
-    def test_s5_11_5_α_v(self):
-        '''
-        Test the s5_11_5_α_v method.
-        '''
-
-        self.fail()
-
-    def test_s5_11_3_Non_uniform_shear_factor(self):
-        '''
-        Test the s5_11_3_Non_uniform_shear_factor method.
-        '''
-
-        self.fail()
-
-    def test_s5_11_2_V_u(self):
-        '''
-        Test the s5_11_2_V_u method.
-        '''
-
-        self.fail()
