+++ AS4100 Modules/Verification/UT_AS4100_5_Shear.py
@@ -1,0 +1,32 @@
+from unittest import TestCase
+
+
+class Test_AS4100_5_Shear(TestCase):
+
+    def test_s5_11_4_V_w_Generic(self):
+
+        self.fail()
+
+    def Test_s5_11_4_V_w_WeldInterfaceLimited(self):
+
+        self.fail()
+
+    def Test_s5_11_4_V_w_WeldLimited(self):
+
+        self.fail()
+
+    def Test_s5_11_4_V_w_CHS(self):
+
+        self.fail()
+
+    def Test_s5_11_5_α_v(self):
+
+        self.fail()
+
+    def Test_s5_11_3_Non_uniform_shear_factor(self):
+
+        self.fail()
+
+    def Test_s5_11_2_V_u(self):
+
+        self.fail()
